+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### 功能构建
 
+- 实现预设接收器。
+  - com.dwarfeng.statistics.impl.handler.receiver.DoNothingReceiver。
+
 - 实现预设调度器。
   - com.dwarfeng.statistics.impl.handler.dispatcher.DrainDispatcher。
 
@@ -28,6 +31,7 @@
   - 驱动机制。
   - 调度机制。
   - 主管机制。
+  - 接收机制。
 
 - 定义预设服务。
   - com.dwarfeng.statistics.impl.service.EnabledDriverInfoLookupServiceImpl。